Cambio a ushort

+++ trunk/palsGPU/src/basic/minmin.cu
@@ -12,7 +12,7 @@ void compute_minmin(struct matrix *etc_matrix, struct solution *solution) {
 	// Timming -----------------------------------------------------
 
 	int assigned_tasks[etc_matrix->tasks_count];
-	for (int i = 0; i < etc_matrix->tasks_count; i++) {
+	for (ushort i = 0; i < etc_matrix->tasks_count; i++) {
 		assigned_tasks[i] = 0;
 	}
 
@@ -28,10 +28,10 @@ void compute_minmin(struct matrix *etc_matrix, struct solution *solution) {
 		best_cost = 0.0;
 	
 		// Recorro las tareas.
-		for (int task_i = 0; task_i < etc_matrix->tasks_count; task_i++) {			
+		for (ushort task_i = 0; task_i < etc_matrix->tasks_count; task_i++) {			
 			// Si la tarea task_i no esta asignada.
 			if (assigned_tasks[task_i] == 0) {
-				int best_machine_for_task;
+				ushort best_machine_for_task;
 				best_machine_for_task = 0;
 				
 				float best_machine_cost_for_task;
@@ -39,7 +39,7 @@ void compute_minmin(struct matrix *etc_matrix, struct solution *solution) {
 				best_machine_cost_for_task = solution->machine_compute_time[0] + 
 					get_etc_value(etc_matrix, 0, task_i);
 			
-				for (int machine_x = 1; machine_x < etc_matrix->machines_count; machine_x++) {
+				for (ushort machine_x = 1; machine_x < etc_matrix->machines_count; machine_x++) {
 					float current_cost;
 					
 					current_cost = solution->machine_compute_time[machine_x] + 
@@ -75,7 +75,7 @@ void compute_minmin(struct matrix *etc_matrix, struct solution *solution) {
 	// Actualiza el makespan de la solución.
 	solution->makespan = solution->machine_compute_time[0];
 	
-	for (int i = 1; i < etc_matrix->machines_count; i++) {
+	for (ushort i = 1; i < etc_matrix->machines_count; i++) {
 		if (solution->makespan < solution->machine_compute_time[i]) {
 			solution->makespan = solution->machine_compute_time[i];
 		}
