+++ trunk/palsGPU/src/etc_matrix.h
@@ -11,8 +11,8 @@
 #define ETC_MATRIX_H_
 
 struct matrix {
-	int tasks_count;
-	int machines_count;
+	ushort tasks_count;
+	ushort machines_count;
 	float* data;
 };
 
@@ -20,7 +20,7 @@ struct matrix* create_etc_matrix(struct params *input);
 void free_etc_matrix(struct matrix *etc_matrix);
 void show_etc_matrix(struct matrix *etc_matrix);
 
-void set_etc_value(struct matrix *etc_matrix, int machine, int task, float value);
-float get_etc_value(struct matrix *etc_matrix, int machine, int task);
+void set_etc_value(struct matrix *etc_matrix, ushort machine, ushort task, float value);
+float get_etc_value(struct matrix *etc_matrix, ushort machine, ushort task);
 
 #endif /* ETC_MATRIX_H_ */
